Fix compile in Mingw64.

+++ daemon/wt-monitor-win32.c
@@ -675,10 +675,10 @@ static HANDLE add_watch (SeafWTMonitorPriv *priv,
 
     pthread_mutex_lock (&priv->hash_lock);
     g_hash_table_insert (priv->handle_hash,
-                         g_strdup(repo_id), (gpointer)(long)dir_handle);
+                         g_strdup(repo_id), (gpointer)dir_handle);
 
     info = create_repo_watch_info (repo_id, worktree);
-    g_hash_table_insert (priv->info_hash, (gpointer)(long)dir_handle, info);
+    g_hash_table_insert (priv->info_hash, (gpointer)dir_handle, info);
     pthread_mutex_unlock (&priv->hash_lock);
 
     add_event_to_queue (info->status, WT_EVENT_SCAN_DIR, "", NULL);
